+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for Peppermint
 
+## v0.2.0
+
+  * Added `Peppermint.Connection`. A GenServer that can handle multiple requests simulaniously.
+  * Rewrite to support usage of `Peppermint` in custom processes.
+
 ## v0.1.0
 
 Initial version